Final graident GI fixes to hopefully wrap things up.
Didn't have the spare cycles to test and run. Passing on to public testing.

+++ Gems/AtomLyIntegration/CommonFeatures/Code/Include/AtomLyIntegration/CommonFeatures/GradientGI/GradientGIComponentBus.h
@@ -43,11 +43,13 @@ namespace AZ
             virtual void SetExposure(float exposure) = 0;
             virtual float GetExposure() const = 0;
 
-            // Script Canvas / Lua represent numbers as a signed type, so the scripted
-            // face-resolution API uses int to avoid an unsigned conversion node. Values
-            // are clamped to the valid [4..256] range by the controller.
-            virtual void SetFaceResolution(int resolution) = 0;
-            virtual int GetFaceResolution() const = 0;
+            // Script Canvas / Lua represent every number as a Number (floating point), so the
+            // scripted face-resolution API uses float -- this matches a Number slot directly and
+            // avoids the "expects an integer" implicit-conversion warning node. The value is
+            // rounded and clamped to the valid [4..256] range by the controller. (Getter and
+            // setter must share a type: BehaviorContext VirtualProperty requires it.)
+            virtual void SetFaceResolution(float resolution) = 0;
+            virtual float GetFaceResolution() const = 0;
 
             virtual void SetUpdateMode(GradientGIUpdateMode mode) = 0;
             virtual GradientGIUpdateMode GetUpdateMode() const = 0;
