+++ Gems/AtomLyIntegration/CommonFeatures/Code/Source/GradientGI/GradientGIComponentController.h
@@ -56,8 +56,8 @@ namespace AZ
             void SetGradientColors(const Color& low, const Color& mid, const Color& high) override;
             void SetExposure(float exposure) override;
             float GetExposure() const override;
-            void SetFaceResolution(int resolution) override;
-            int GetFaceResolution() const override;
+            void SetFaceResolution(float resolution) override;
+            float GetFaceResolution() const override;
             void SetUpdateMode(GradientGIUpdateMode mode) override;
             GradientGIUpdateMode GetUpdateMode() const override;
